requirements.txt
requests==2.28.1
apache-airflow[docker,pandas]==2.3.4
xlsxwriter==3.0.3
selenium==4.4.3
pytest==7.1.2